Seurat Enabled Scene named VikingVillage Added
Seurat allows for high quality sceens to be displayed in mobile vr headsets allowing for free 6DoF head movement in a small space. In this case in the area between wooden door frame and beginning of the bridge.
You can test it by signing and installing the VikingVillage.apk

If you wanne build the project yourself:
(Some files needed to be compressed for upload please look through the project and unsip them if you wanna open and run the project in unity)
Big thank you to reddit user st6315 for giving me access to his scene for free.