Daydream/Cardboard version updated to ARCore 1.5.0.
Daydream/Cardboard version updated to ARCore 1.5.0.
Implemented Tivoyagefeaks suggestion together with latest updates. Unfortenetly I could not confirm that it works. But since it doesn't effect the performance and might work when using the controller I added it and Tivoyagefeaks might be able to fix it quickly when I missed something that it prevents it from working. I renamed the class to AlignmentCorrection and added the mirrorAngle variable to the FollowARCoreCamera class.

#17