Update to ARCore 1.6, distortions during head rotations fixed, head tracking speed and smoothness optimized, Moving Marker Tracking Project added (proof of concept)
The update to ARCore 1.6 brings supports for more devices but this is only relevant for the Charboard/Daydream versions of this app which isn't ready yet.

Head tracking was further Improved by doing a deeper analysis of the underling problem and choosing better settings and a new interpolation method.
The underling problem is that the ARCore tracking jumps frequently and the jumps need to be smoothed out.
The current version tries to find the right balance between lag and responsivness. A smoother experience is possible but it also introduces very noticable lag. If interested in a smoother head movement please set the settings of the SmoothDamp method in the followARCamera script to higer values.

By moving the user camera behind the ar camera the distorition during head rotations could be eliminated.

Moving marker tracking project added. This is still a proof of concept in development and far from perfect but it is possible to track a moving marker while beeing in vr and also use ARCore for head tracking at the same time. The marker that is recognised by the camera is the king.png which can be found in the MovingMarkerTracking project file.