2 experiementel GearVR scenes and apps added. Solved the lowerCase issue of some script names
TDLR: The 2 experimental tracking apps might deliver better tracking on S8, S9 and Note8.

The 2 experimental scenes might deliver better tracking responsivenes on modern GearVR devices like S8, Note8 or S9 but do not work well on the S7.
Booth experimental apps should deliver roughly the same experience the difference is in the Unity implementation of ARCore that I used to build them. I think there are differences in the performence between the two apps and I will compare them in detail soon.
They are marked as experimental since they do not work well on the S7 but I believe they could work very well on modern devices like S8, S9, Note8.

If you have one of the newer devices and a GearVR give them a try and let me know if they work well.

Some Backround:
1) The S7 is not compatible with Vuforia Fusion that's why the FuforiaTracking app only works on modern devices.

2) The activation of ARCore and VR in one app produces some unwanted behaivior which costs an enormus amount of performance. The NativeARCoreTracking app has therefore high performance requirements which the S7 can not handel. Newer devices have a bit more power and might be able to hit the 60 fps and therfore deliver smoother tracking.

Fuvoria allows the developer to activate VR and ARCore in one app and I want to find out if that helps to eliminate the unwanted behaivior we see in the NativeARCoreTracking app. If it does we would have more performance left for doing other stuff.

If you have time and interest give the experimental apps a try and let me know if they work smooth and solid or if the tracking jitters like crazy.

I also fixed the issue of some methods throwing errors since the capital letter was not updated by github automatically. fixes #21